+++ CHANGES.md
@@ -13,6 +13,7 @@ versions and prefixed with the kind of change.
     - UPGRADE: upgrade to Vue 3.2.32
     - UPGRADE: upgrade to ESLint 18
     - BUGFIX: fix audio handling in case of 0 channels
+    - BUGFIX: try to stop FFmpeg sub-process in a more clean way
     - CLEANUP: make ESLint happy again
 
 - 2.8.0